Serverlock: colocate query structs
